refactor: consolidate telegram_bot/ into telegram/ package and integrate TurnExecutor
- Rename telegram_bot/ to telegram/ package
- Add TurnExecutor and Conversation to sdk/turn for channel abstraction
- Add 'model' config field to TelegramBotConfig
- Refactor message_handler to use TurnExecutor
- Update aiohttp_app to use TelegramChannel
- Update tests and dependencies
