+++ server/message_handler.py
@@ -1,69 +1,34 @@
 """Message handler for user prompts."""
 
-import asyncio
 import logging
-from collections.abc import AsyncGenerator, Callable, Sequence
-from contextlib import suppress
-from dataclasses import dataclass, field
+from collections.abc import AsyncGenerator, Sequence
 
-from sdk.context import ContextManager, ConversationHistory, LLMCompactionStrategy, ToolClearingStrategy
-from sdk.events import (
-    AgentEvent,
-    ContentPayload,
-    TurnEndPayload,
-    agent_span,
-    get_current_dispatcher,
-    set_model_options,
-)
-from sdk.turn import turn_scope
-from agents.types import Agent, Data, LLMOptions
-from tools.memory import load_memory
+from sdk.context import ConversationHistory
+from sdk.events import AgentEvent, ContentPayload, TurnEndPayload
+from sdk.turn import Conversation, TurnExecutor
+from agents.types import Data, LLMOptions
 from tools.virtual_computer.receive_file import receive_attachment
 
-from agents import AVAILABLE_AGENTS, resolve_agent as _resolve_agent
-from conversations import (
-    generate_conversation_title,
-    load_conversation_history,
-    load_loaded_skills,
-    save_agent_events,
-    save_conversation_title,
-    save_loaded_skills,
-)
-from sdk import (
-    PersistenceHook,
-    default_hooks,
-    run_turn,
-)
-from sdk.hooks._agent_event_buffer import AgentEventBufferHook
-from sdk.skills.agent_state import AgentState
-from sdk.tools._core import get_core_tools
-from sdk.turn._turn import StopRequestedError
+from agents import AVAILABLE_AGENTS
+from conversations import load_conversation_history
 
 logger = logging.getLogger(__name__)
 
 _DEFAULT_CONVERSATION_ID = "default"
 
-
-@dataclass
-class _Conversation:
-    """Per-conversation state: conversation history and context manager."""
-
-    history: ConversationHistory = field(default_factory=ConversationHistory)
-    context_manager: ContextManager | None = None
-
+# Module-level turn executor shared across all SSE conversations.
+_turn_executor = TurnExecutor()
 
 # Conversation store keyed by conversation ID.
-_conversations: dict[str, _Conversation] = {}
-
-# Track background tasks to avoid garbage collection (RUF006)
-_background_tasks: set[asyncio.Task] = set()
+_conversations: dict[str, Conversation] = {}
 
 
-def _get_conversation(conversation_id: str | None = None) -> _Conversation:
+def _get_conversation(conversation_id: str | None = None) -> Conversation:
     """Return the conversation for the given ID, creating one if needed."""
     cid = conversation_id or _DEFAULT_CONVERSATION_ID
     if cid not in _conversations:
-        _conversations[cid] = _Conversation(
+        _conversations[cid] = Conversation(
+            id=cid,
             history=ConversationHistory(instance_id=cid),
         )
     return _conversations[cid]
@@ -84,28 +49,12 @@ def resume_conversation(conversation_id: str) -> list[dict] | None:
     if messages is None:
         return None
 
-    conversation = _Conversation(
+    conversation = Conversation(
+        id=conversation_id,
         history=ConversationHistory(messages, instance_id=conversation_id),
     )
     _conversations[conversation_id] = conversation
     return messages
-
-
-def _refresh_system_message(history: ConversationHistory, system_prompt: str) -> None:
-    """Re-inserts the system message at the start of history with up-to-date memory.
-
-    Called before each model invocation so any memories stored during the previous
-    turn are visible immediately.
-    """
-    instruction = system_prompt
-    memory = load_memory()
-    if memory:
-        lines = "\n".join(f"  {k}: {e.value}" for k, e in memory.items())
-        sep = "─" * 64
-        memory_block = f"\n── Memory (persisted across sessions) ──────────────────────────\n{lines}\n{sep}\n"
-        instruction = memory_block + instruction
-
-    history.set_system_message(instruction)
 
 
 def _augment_message_with_attachments(message: str, data: Sequence[Data]) -> str:
@@ -126,132 +75,6 @@ def _augment_message_with_attachments(message: str, data: Sequence[Data]) -> str
         f"[Attached files written to virtual computer]\n"
         f"{files_block}"
     )
-
-
-def _build_agent(agent_id: str | None, options: LLMOptions) -> Agent:
-    """Construct an Agent from a registry ID and LLM options."""
-    agent_name, agent_desc, agent_prompt, agent_tools = _resolve_agent(agent_id)
-    return Agent(
-        name=agent_name,
-        description=agent_desc,
-        instruction=agent_prompt,
-        tools=agent_tools,
-        model=options.model,  # type: ignore[arg-type]  # validated by caller
-        think=options.think or False,
-        persist_thinking=options.persist_thinking if options.persist_thinking is not None else True,
-        options=options.to_options(),
-        max_iterations=options.max_iterations or 0,
-    )
-
-
-def _ensure_context_manager(
-    conversation: _Conversation,
-    active_agent: Agent,
-) -> ContextManager:
-    """Return the conversation's context manager, creating it if needed."""
-    if conversation.context_manager is None:
-        num_ctx = active_agent.options.get("num_ctx", 0) if active_agent.options else 0
-        conversation.context_manager = ContextManager(
-            history=conversation.history,
-            context_limit=num_ctx,
-            agent_name=active_agent.name,
-            strategies=[ToolClearingStrategy(), LLMCompactionStrategy()],
-        )
-    return conversation.context_manager
-
-
-async def _run_turn(
-    *,
-    conversation: _Conversation,
-    active_agent: Agent,
-    user_content: str,
-    options: LLMOptions,
-    conversation_id: str | None,
-    handler: Callable[[AgentEvent], object],
-    is_new_conversation: bool = False,
-) -> None:
-    """Execute a single conversation turn: model calls, tool execution, persistence."""
-    logger.info(
-        "Turn started: conv=%s agent=%s message=%.80s",
-        conversation_id or _DEFAULT_CONVERSATION_ID,
-        active_agent.name,
-        user_content,
-    )
-    set_model_options(options)
-
-    ctx_manager = _ensure_context_manager(conversation, active_agent)
-    conv_id = conversation_id or _DEFAULT_CONVERSATION_ID
-
-    # Fresh AgentState each turn, restored from persisted skill names.
-    agent_state = AgentState(get_core_tools() + active_agent.tools)
-    for skill_name in load_loaded_skills(conv_id):
-        if agent_state.load(skill_name) is not None:
-            logger.info("Restored skill '%s' for conv=%s", skill_name, conv_id)
-
-    async with turn_scope(handler=handler, conversation_id=conversation_id):
-        # Subscribe event buffer to capture agent lifecycle/preview events
-        event_buffer = AgentEventBufferHook()
-        dispatcher = get_current_dispatcher()
-        if dispatcher:
-            dispatcher.subscribe(event_buffer.handle_event)
-
-        with agent_span(active_agent.name, instruction=user_content, agent_state=agent_state):
-            conversation.history.append({"role": "user", "content": user_content})
-            _refresh_system_message(conversation.history, active_agent.instruction)
-
-            hooks = default_hooks(
-                active_agent,
-                max_iterations=active_agent.max_iterations,
-                ctx_manager=ctx_manager,
-            )
-
-            hooks.append(PersistenceHook(
-                conversation_id=conv_id,
-                history=conversation.history,
-            ))
-
-            with suppress(StopRequestedError):
-                await run_turn(
-                    history=conversation.history,
-                    agent=active_agent,
-                    hooks=hooks,
-                )
-
-        # Persist loaded skills so they survive across turns and restarts
-        if agent_state.loaded_skill_names:
-            try:
-                save_loaded_skills(conv_id, agent_state.loaded_skill_names)
-            except Exception:
-                logger.exception("Failed to save loaded skills for '%s'", conv_id)
-
-        # Yield to event loop so call_soon callbacks (sync event handlers)
-        # have a chance to run before we read the buffer
-        await asyncio.sleep(0)
-
-        # Save agent events after the turn (outside agent_span so completion is captured)
-        buffered_events = event_buffer.get_events()
-        if buffered_events:
-            try:
-                save_agent_events(conv_id, buffered_events)
-                logger.info("Saved %d agent events for conv=%s", len(buffered_events), conv_id)
-            except Exception:
-                logger.exception("Failed to save agent events for '%s'", conv_id)
-
-        # Generate a title for new conversations after the first successful turn
-        if is_new_conversation and conversation_id:
-            task = asyncio.create_task(_generate_title(conversation_id, user_content))
-            _background_tasks.add(task)
-            task.add_done_callback(_background_tasks.discard)
-
-
-async def _generate_title(conversation_id: str, first_message: str) -> None:
-    """Generate and save a title for a new conversation."""
-    try:
-        title = await generate_conversation_title(first_message)
-        save_conversation_title(conversation_id, title)
-        logger.info("Generated title for conversation %s: %r", conversation_id, title)
-    except Exception:
-        logger.exception("Failed to generate title for conversation %s", conversation_id)
 
 
 async def handle_user_message(
@@ -287,44 +110,14 @@ async def handle_user_message(
         raise ValueError(msg)
 
     try:
-        # Bridge published events via a queue so we can stream them to the caller.
-        queue: asyncio.Queue[AgentEvent | None] = asyncio.Queue()
-
-        async def _queue_handler(evt: AgentEvent) -> None:
-            try:
-                await queue.put(evt)
-            except Exception:  # pragma: no cover - defensive logging
-                logger.exception("Failed to enqueue AgentEvent in message handler")
-
-        active_agent = _build_agent(agent, options)
-
-        async def _producer() -> None:
-            try:
-                await _run_turn(
-                    conversation=conversation,
-                    active_agent=active_agent,
-                    user_content=user_content,
-                    options=options,
-                    conversation_id=conversation_id,
-                    handler=_queue_handler,
-                    is_new_conversation=is_new_conversation,
-                )
-            finally:
-                await queue.put(None)
-
-        producer_task = asyncio.create_task(_producer())
-        try:
-            while True:
-                item = await queue.get()
-                if item is None:
-                    break
-                yield item
-        finally:
-            if not producer_task.done():
-                producer_task.cancel()
-            with suppress(Exception):
-                await producer_task
-
+        async for event in _turn_executor.execute(
+            conversation=conversation,
+            user_content=user_content,
+            agent_id=agent,
+            options=options,
+            is_new_conversation=is_new_conversation,
+        ):
+            yield event
     except Exception:
         logger.exception("Error handling user message")
         yield AgentEvent(payload=ContentPayload(
