+++ tests/sdk/events/test_message_handler_bridge.py
@@ -49,9 +49,9 @@ async def test_message_handler_bridges_events_without_duplicates(monkeypatch: py
 
         return "done"
 
-    import server.message_handler as mh
+    import sdk.turn._executor as _executor_mod
 
-    monkeypatch.setattr(mh, "run_turn", _fake_tool_loop)
+    monkeypatch.setattr(_executor_mod, "run_turn", _fake_tool_loop)
     seen: list[AgentEvent] = []
     async for ev in handle_user_message("hi", data=None, options=LLMOptions(model="test-model")):
         seen.append(ev)
